+++ yash-syntax/src/parser/lex/tilde.rs
@@ -30,14 +30,15 @@ use crate::syntax::WordUnit::{self, Tilde, Unquoted};
 /// slash character (or a colon character if `delimit_at_colon` is `true`).
 ///
 /// If successful, this function returns a tuple of the length of the parsed
-/// word units (including the tilde character) and the name of the tilde
-/// expansion (excluding the tilde character and the delimiter). Note that the
-/// name may be empty.
+/// word units (including the tilde character), the name of the tilde
+/// expansion (excluding the tilde character and the delimiter), and a Boolean
+/// indicating whether the name is followed by a slash character.
+/// Note that the name may be empty.
 ///
 /// If the first word unit is not an unquoted tilde character or the name is
 /// delimited by a word unit other than an unquoted literal character, this
 /// function returns `None`.
-fn parse_tilde<'a, I>(units: I, delimit_at_colon: bool) -> Option<(usize, String)>
+fn parse_tilde<'a, I>(units: I, delimit_at_colon: bool) -> Option<(usize, String, bool)>
 where
     I: IntoIterator<Item = &'a WordUnit>,
 {
@@ -51,7 +52,7 @@ where
 
     for unit in units {
         match unit {
-            Unquoted(Literal('/')) => break,
+            Unquoted(Literal('/')) => return Some((count, name, true)),
             Unquoted(Literal(':')) if delimit_at_colon => break,
             Unquoted(Literal(c)) => {
                 name.push(*c);
@@ -61,7 +62,7 @@ where
         }
     }
 
-    Some((count, name))
+    Some((count, name, false))
 }
 
 impl Word {
@@ -78,7 +79,7 @@ impl Word {
     /// # use yash_syntax::syntax::{Word, WordUnit::Tilde};
     /// let mut word = Word::from_str("~").unwrap();
     /// word.parse_tilde_front();
-    /// assert_eq!(word.units, [Tilde("".to_string())]);
+    /// assert_eq!(word.units, [Tilde { name: "".to_string(), followed_by_slash: false }]);
     /// ```
     ///
     /// ```
@@ -86,7 +87,7 @@ impl Word {
     /// # use yash_syntax::syntax::{Word, WordUnit::Tilde};
     /// let mut word = Word::from_str("~foo").unwrap();
     /// word.parse_tilde_front();
-    /// assert_eq!(word.units, [Tilde("foo".to_string())]);
+    /// assert_eq!(word.units, [Tilde { name: "foo".to_string(), followed_by_slash: false }]);
     /// ```
     ///
     /// If there is no leading tilde, `self.units` will have the same content
@@ -115,8 +116,14 @@ impl Word {
     /// delimiters.
     #[inline]
     pub fn parse_tilde_front(&mut self) {
-        if let Some((len, name)) = parse_tilde(&self.units, false) {
-            self.units.splice(..len, std::iter::once(Tilde(name)));
+        if let Some((len, name, followed_by_slash)) = parse_tilde(&self.units, false) {
+            self.units.splice(
+                ..len,
+                std::iter::once(Tilde {
+                    name,
+                    followed_by_slash,
+                }),
+            );
         }
     }
 
@@ -135,13 +142,13 @@ impl Word {
     /// assert_eq!(
     ///     word.units,
     ///     [
-    ///         Tilde("".to_string()),
+    ///         Tilde { name: "".to_string(), followed_by_slash: false },
     ///         Unquoted(Literal(':')),
-    ///         Tilde("a".to_string()),
+    ///         Tilde { name: "a".to_string(), followed_by_slash: true },
     ///         Unquoted(Literal('/')),
     ///         Unquoted(Literal('b')),
     ///         Unquoted(Literal(':')),
-    ///         Tilde("c".to_string()),
+    ///         Tilde { name: "c".to_string(), followed_by_slash: false },
     ///     ]
     /// );
     /// ```
@@ -175,11 +182,11 @@ impl Word {
     ///         Unquoted(Literal('=')),
     ///         // This tilde is parsed because it is at index 2,
     ///         // even though it is not after a colon.
-    ///         Tilde("a".to_string()),
+    ///         Tilde { name: "a".to_string(), followed_by_slash: true },
     ///         Unquoted(Literal('/')),
     ///         Unquoted(Literal('b')),
     ///         Unquoted(Literal(':')),
-    ///         Tilde("c".to_string()),
+    ///         Tilde { name: "c".to_string(), followed_by_slash: false },
     ///     ]
     /// );
     /// ```
@@ -190,8 +197,14 @@ impl Word {
         let mut i = index;
         loop {
             // Parse a tilde expansion at index `i`.
-            if let Some((len, name)) = parse_tilde(&self.units[i..], true) {
-                self.units.splice(i..i + len, std::iter::once(Tilde(name)));
+            if let Some((len, name, followed_by_slash)) = parse_tilde(&self.units[i..], true) {
+                self.units.splice(
+                    i..i + len,
+                    std::iter::once(Tilde {
+                        name,
+                        followed_by_slash,
+                    }),
+                );
                 i += 1;
             }
 
@@ -247,7 +260,13 @@ mod tests {
         let input = Word::from_str("~").unwrap();
         let result = parse_tilde_front(&input);
         assert_eq!(result.location, input.location);
-        assert_eq!(result.units, [Tilde("".to_string())]);
+        assert_eq!(
+            result.units,
+            [Tilde {
+                name: "".to_string(),
+                followed_by_slash: false
+            }]
+        );
     }
 
     #[test]
@@ -255,7 +274,13 @@ mod tests {
         let input = Word::from_str("~foo").unwrap();
         let result = parse_tilde_front(&input);
         assert_eq!(result.location, input.location);
-        assert_eq!(result.units, [Tilde("foo".to_string())]);
+        assert_eq!(
+            result.units,
+            [Tilde {
+                name: "foo".to_string(),
+                followed_by_slash: false
+            }]
+        );
     }
 
     #[test]
@@ -266,7 +291,10 @@ mod tests {
         assert_eq!(
             result.units,
             [
-                Tilde("bar".to_string()),
+                Tilde {
+                    name: "bar".to_string(),
+                    followed_by_slash: true,
+                },
                 Unquoted(Literal('/')),
                 SingleQuote("".to_string()),
             ]
@@ -278,7 +306,13 @@ mod tests {
         let input = Word::from_str("~bar:baz").unwrap();
         let result = parse_tilde_front(&input);
         assert_eq!(result.location, input.location);
-        assert_eq!(result.units, [Tilde("bar:baz".to_string())]);
+        assert_eq!(
+            result.units,
+            [Tilde {
+                name: "bar:baz".to_string(),
+                followed_by_slash: false
+            }]
+        );
     }
 
     #[test]
@@ -364,7 +398,10 @@ mod tests {
         assert_eq!(
             result.units,
             [
-                Tilde("a".to_string()),
+                Tilde {
+                    name: "a".to_string(),
+                    followed_by_slash: true,
+                },
                 Unquoted(Literal('/')),
                 Unquoted(Literal('b')),
                 Unquoted(Literal(':')),
@@ -396,7 +433,13 @@ mod tests {
         let input = Word::from_str("~").unwrap();
         let result = parse_tilde_everywhere(&input);
         assert_eq!(result.location, input.location);
-        assert_eq!(result.units, [Tilde("".to_string())]);
+        assert_eq!(
+            result.units,
+            [Tilde {
+                name: "".to_string(),
+                followed_by_slash: false
+            }]
+        );
     }
 
     #[test]
@@ -404,7 +447,13 @@ mod tests {
         let input = Word::from_str("~foo").unwrap();
         let result = parse_tilde_everywhere(&input);
         assert_eq!(result.location, input.location);
-        assert_eq!(result.units, [Tilde("foo".to_string())]);
+        assert_eq!(
+            result.units,
+            [Tilde {
+                name: "foo".to_string(),
+                followed_by_slash: false
+            }]
+        );
     }
 
     #[test]
@@ -415,7 +464,10 @@ mod tests {
         assert_eq!(
             result.units,
             [
-                Tilde("bar".to_string()),
+                Tilde {
+                    name: "bar".to_string(),
+                    followed_by_slash: true
+                },
                 Unquoted(Literal('/')),
                 SingleQuote("".to_string()),
             ]
@@ -430,7 +482,10 @@ mod tests {
         assert_eq!(
             result.units,
             [
-                Tilde("bar".to_string()),
+                Tilde {
+                    name: "bar".to_string(),
+                    followed_by_slash: false
+                },
                 Unquoted(Literal(':')),
                 DoubleQuote(Text(vec![])),
             ]
@@ -490,7 +545,13 @@ mod tests {
         assert_eq!(result.location, input.location);
         assert_eq!(
             result.units,
-            [Unquoted(Literal(':')), Tilde("".to_string())]
+            [
+                Unquoted(Literal(':')),
+                Tilde {
+                    name: "".to_string(),
+                    followed_by_slash: false
+                }
+            ]
         );
 
         let input = Word::from_str(":~foo/a:~bar").unwrap();
@@ -500,11 +561,17 @@ mod tests {
             result.units,
             [
                 Unquoted(Literal(':')),
-                Tilde("foo".to_string()),
+                Tilde {
+                    name: "foo".to_string(),
+                    followed_by_slash: true,
+                },
                 Unquoted(Literal('/')),
                 Unquoted(Literal('a')),
                 Unquoted(Literal(':')),
-                Tilde("bar".to_string()),
+                Tilde {
+                    name: "bar".to_string(),
+                    followed_by_slash: false
+                },
             ]
         );
 
@@ -514,16 +581,25 @@ mod tests {
         assert_eq!(
             result.units,
             [
-                Tilde("a".to_string()),
+                Tilde {
+                    name: "a".to_string(),
+                    followed_by_slash: true,
+                },
                 Unquoted(Literal('/')),
                 Unquoted(Literal('b')),
                 Unquoted(Literal(':')),
-                Tilde("c".to_string()),
+                Tilde {
+                    name: "c".to_string(),
+                    followed_by_slash: true,
+                },
                 Unquoted(Literal('/')),
                 Unquoted(Literal('d')),
                 Unquoted(Literal(':')),
                 Unquoted(Literal(':')),
-                Tilde("".to_string()),
+                Tilde {
+                    name: "".to_string(),
+                    followed_by_slash: false
+                },
             ]
         );
     }