+++ yash-semantics/CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- When a tilde expansion produces a directory name that ends with a slash and
+  the expansion is followed by a slash, the trailing slash in the directory name
+  is now removed to maintain the correct number of slashes.
+    - This is done by removing the trailing slash from the directory name in
+      `impl expansion::initial::Expand for WordUnit`. This is done only if the
+      `followed_by_slash` flag is set to `true` in the tilde expansion
+      (`WordUnit::Tilde`).
 - External dependency versions:
     - yash-env 0.7.1 → 0.8.0
     - yash-syntax 0.14.1 → 0.15.0