+++ yash-semantics/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to `yash-semantics` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - Unreleased
+
+### Changed
+
+- When a tilde expansion produces a directory name that ends with a slash and
+  the expansion is followed by a slash, the trailing slash in the directory name
+  is now removed to maintain the correct number of slashes.
+    - This is done by removing the trailing slash from the directory name in
+      `impl expansion::initial::Expand for WordUnit`. This is done only if the
+      `followed_by_slash` flag is set to `true` in the tilde expansion
+      (`WordUnit::Tilde`).
+- External dependency versions:
+    - yash-env 0.7.1 → 0.8.0
+    - yash-syntax 0.14.1 → 0.15.0
+
 ## [0.7.1] - 2025-05-03
 
 ### Changed
@@ -229,6 +244,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of the `yash-semantics` crate
 
+[0.8.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.8.0
 [0.7.1]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.7.1
 [0.7.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.7.0
 [0.6.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.6.0